vit/requirements.txt
torch
torchvision
timm
matplotlib
numpy
scikit-learn
tensorboard
pyyaml
